+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Next
 
+## 3.8.3 - 2024-10-25
+
+- recording: fix crash when calling view.isVisible ([#201](https://github.com/PostHog/posthog-android/pull/201))
+- recording: change debouncerDelayMs default from 500ms to 1000ms (1s) ([#201](https://github.com/PostHog/posthog-android/pull/201))
+
 ## 3.8.2 - 2024-10-14
 
 - recording: session replay respect feature flag variants ([#197](https://github.com/PostHog/posthog-android/pull/197))